test: adds test to assert graceful shutdown works
Documentation about this feature is [here](https://kubernetes.io/docs/concepts/cluster-administration/node-shutdown/)

Note that enabling this only works when the node is initialized the
first time. It looks as though the `kubelet.yaml` configuration file is
only created when the worker is created initially, so modifying the
values in `workerProfile` is not reflected until a new worker is created
for that profile.

For this reason, it might be a good idea to either turn this on by
default or make it more prominent in the config like `k3s` does
[here](https://github.com/NixOS/nixpkgs/blob/1d3aeb5a193b9ff13f63f4d9cc169fb88129f860/nixos/modules/services/cluster/k3s/default.nix#L391)