feat: specify nix formatter to use
This specifies `nixfmt-tree` as the nix formatter to use when invoking `nix
fmt`.

`nixfmt-tree` is a wrapper for both [`nixfmt`](https://github.com/NixOS/nixfmt)
(the official nix formatter) and [`treefmt`](https://github.com/numtide/treefmt)
which will run a formatter over a tree of files, providing good defaults like
respecting the `.gitinore` file.

This updates the nixpks url to the latest 25.05 relase so we can get the
latest `nixfmt` packages.

The end result is that to format the files in this flake, you can simply run
`nix fmt`.

It also adds a github action check to verify that the files are
formatted. The check will fail until the files are formatted, after
which we can add an entry to `.git-blame-ignore-revs` to ignore that
commit during `git blame`s.