Calico add hostNetwork option
This adds the `hostNetwork` option to the Calico module.

The Calico CNI can get into a startup deadlock where the `calico-node`
pod fails to start because its CNI network can't be set up. This
happens because the CNI plugin itself needs a `/var/lib/calico/nodename`
file, which the `calico-node` pod is supposed to create.

Setting `hostNetwork: true` on the `calico-node` DaemonSet is the
standard upstream solution to break this cycle.

This change exposes that option so users can configure Calico correctly
and avoid this race condition declaratively.